requirements.txt
arrow==0.5.0
biopython==1.70
certifi==2018.1.18
Flask==0.10.1
future==0.14.3
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
memoized-property==1.0.2
mendeley==0.3.2
numpy==1.14.1
oauthlib==0.7.2
pyflakes==1.1.0
python-dateutil==2.4.0
PyYAML==3.11
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
Werkzeug==0.10
wincertstore==0.2
